Add option for AST debug information on hover
