+++ package.json
@@ -52,19 +52,19 @@
                     "scope": "resource",
                     "type": "boolean",
                     "default": true,
-                    "description": "Uses the legacy OCaml language server when `dfx <= 0.11.0`."
+                    "description": "Use the legacy OCaml language server when `dfx <= 0.11.0`."
                 },
                 "motoko.maxNumberOfProblems": {
                     "scope": "resource",
                     "type": "integer",
                     "default": 100,
-                    "description": "Limits the number of warnings and errors generated by the compiler."
+                    "description": "Limit the number of warnings and errors generated by the compiler."
                 },
                 "motoko.hideWarningRegex": {
                     "scope": "resource",
                     "type": "string",
                     "default": "",
-                    "description": "Hides compiler warnings based on a regular expression."
+                    "description": "Hide compiler warnings based on a regular expression."
                 },
                 "motoko.trace.server": {
                     "scope": "window",
@@ -75,7 +75,11 @@
                         "verbose"
                     ],
                     "default": "off",
-                    "description": "Traces the communication between VS Code and the language server."
+                    "description": "Trace the communication between VS Code and the language server."
+                },
+                "motoko.debugHover": {
+                    "type": "boolean",
+                    "description": "Show debug information when hovering over any part of a Motoko program."
                 },
                 "motoko.canister": {
                     "scope": "resource",
